Fixed favorites and added README

+++ src/components/RecipeCard.js
@@ -1,6 +1,10 @@
 import { Link } from "react-router-dom";
+import { toast } from "react-toastify";
 
-const RecipeCard = ({ meal }) => {
+const RecipeCard = ({
+  meal,
+  showFavoriteButton = true,
+}) => {
 
   const addToFavorites = () => {
 
@@ -9,6 +13,17 @@ const RecipeCard = ({ meal }) => {
         localStorage.getItem("favorites")
       ) || [];
 
+    const exists = favorites.some(
+      (item) => item.idMeal === meal.idMeal
+    );
+
+    if (exists) {
+      toast.info(
+        "Already in Favorites ❤️"
+      );
+      return;
+    }
+
     favorites.push(meal);
 
     localStorage.setItem(
@@ -16,11 +31,17 @@ const RecipeCard = ({ meal }) => {
       JSON.stringify(favorites)
     );
 
-    alert("Added to Favorites ❤️");
+    window.dispatchEvent(
+      new Event("favoritesUpdated")
+    );
+
+    toast.success(
+      "Added to Favorites ❤️"
+    );
   };
 
   return (
-    <div className="bg-white/10 backdrop-blur-lg rounded-2xl overflow-hidden shadow-xl border border-white/10 hover:scale-105 hover:shadow-pink-500/30 transition duration-300">
+    <div className="bg-white/10 backdrop-blur-lg rounded-2xl overflow-hidden shadow-xl border border-white/10 hover:scale-105 transition duration-300">
 
       <Link to={`/recipe/${meal.idMeal}`}>
 
@@ -44,12 +65,14 @@ const RecipeCard = ({ meal }) => {
 
       </Link>
 
-      <button
-        onClick={addToFavorites}
-        className="bg-gradient-to-r from-pink-500 to-red-500 hover:from-yellow-400 hover:to-pink-500 text-white w-full py-3 font-bold transition duration-300 hover:bg-pink-600 text-white w-full py-3 font-bold transition"
-      >
-        ❤️ Add to Favorites
-      </button>
+      {showFavoriteButton && (
+        <button
+          onClick={addToFavorites}
+          className="bg-gradient-to-r from-pink-500 to-red-500 text-white w-full py-3 font-bold"
+        >
+          ❤️ Add to Favorites
+        </button>
+      )}
 
     </div>
   );
